code_review_agent/requirements.txt
openai>=1.0.0
pyyaml>=6.0
requests>=2.31.0
anthropic>=0.19.0
click>=8.1.0